[refactor][#1] 코드별 파일 분리
